+++ resources/lang/en/resource-type-item/parameters.php
@@ -7,5 +7,5 @@ return [
     'description-include-categories' => 'Optionally include the category assigned to each item',
 
     'title-include-subcategories' => 'Include assigned subcategory?',
-    'description-include-subcategories' => 'Optionally include the subcategory assigned to each item'
+    'description-include-subcategories' => 'Optionally include the subcategory assigned to each item, include-categories must be set'
 ];